:building_construction: (477) organise evaluation codes with distinct folders in the evaluation folder
