+++ mcr-core/mcr_meeting/evaluation/asr/utils/metrics_calculator.py
@@ -10,7 +10,7 @@ from pyannote.metrics.diarization import (
 from mcr_meeting.app.schemas.transcription_schema import (
     DiarizedTranscriptionSegment,
 )
-from mcr_meeting.evaluation.eval_types import (
+from mcr_meeting.evaluation.asr.types import (
     DiarizationMetrics,
     EvaluationMetrics,
     MetricsPipelineInput,
