+++ mcr-core/mcr_meeting/evaluation/asr/utils/result_manager.py
@@ -5,7 +5,7 @@ import pandas as pd
 from loguru import logger
 
 from mcr_meeting.app.services.s3_service import put_file_to_s3, s3_settings
-from mcr_meeting.evaluation.eval_types import (
+from mcr_meeting.evaluation.asr.types import (
     EvaluationOutput,
     EvaluationSummary,
     TranscriptionOutput,
