+++ mcr-core/mcr_meeting/evaluation/utils/__init__.py
@@ -1,4 +1,4 @@
-from mcr_meeting.evaluation.utils.metrics_calculator import MetricsCalculator
-from mcr_meeting.evaluation.utils.result_manager import ResultsManager
+from mcr_meeting.evaluation.asr.utils.metrics_calculator import MetricsCalculator
+from mcr_meeting.evaluation.asr.utils.result_manager import ResultsManager
 
 __all__ = ["MetricsCalculator", "ResultsManager"]
